Handle both kinds of python-magic
Unfortunately there's two different libraries called python-magic, and they both
do the same thing. To make matters more complex, it's easy to find either
versions on most systems. For instance; Fedora 23 comes with a different version
of python-magic then the one included on pip. This makes it impossible to
actually use the version from pip instead because various RPMs rely on this
version of python-magic.